+++ code/packages/create-frontend-app/template/+nextjs+websockets/components/home/home.tsx
@@ -6,6 +6,7 @@ import { motion } from 'motion/react';
 import { Logo } from '@/components/utils/logo';
 import LoginForm from '../home-components/login-form';
 import Dither from '../ui/reactbits/Backgrounds/Dither/Dither';
+// import RotatingText from "../ui/reactbits/TextAnimations/RotatingText/RotatingText";
 import { Button } from '../ui/button';
 import { Github, Book } from 'lucide-react';
 import { DOCUMENTATION_URL, GITHUB_URL } from '@/lib/constants';
@@ -35,7 +36,7 @@ export const Home = () => {
         >
           <div className="max-w-[520px] w-full flex flex-col items-center justify-between gap-0">
             <div className="w-full flex justify-between items-center gap-2 md:left-8 md:top-8 bg-background p-8 py-6 border border-[#c9c9c9]">
-              <Logo />
+              <Logo kind="small" />
               <motion.div
                 initial={{ opacity: 0, y: -20 }}
                 animate={{ opacity: 1, y: 0 }}
@@ -43,10 +44,34 @@ export const Home = () => {
                 className="flex flex-col items-end justify-center"
               >
                 <h1 className="text-3xl font-inter font-bold text-foreground uppercase">
-                  BOILERPLATE
+                  WHITEBOARD
                 </h1>
+                <h2 className="text-2xl font-inter font-light text-muted-foreground uppercase">
+                  SHOWCASE
+                </h2>
               </motion.div>
             </div>
+            {/* <div className="w-full flex gap-1 items-center justify-center bg-transparent">
+              <RotatingText
+                texts={[
+                  "collaborative",
+                  "easy to use",
+                  "extensible",
+                  "visual",
+                  "open source",
+                ]}
+                mainClassName="font-inter font-light leading-[100px] w-full h-full text-black overflow-hidden justify-center items-center text-3xl"
+                staggerFrom={"last"}
+                initial={{ y: "100%" }}
+                animate={{ y: 0 }}
+                exit={{ y: "-120%" }}
+                staggerDuration={0.025}
+                splitBy="characters"
+                splitLevelClassName="overflow-hidden pb-0.5 sm:pb-1 md:pb-1"
+                transition={{ type: "spring", damping: 30, stiffness: 400 }}
+                rotationInterval={2000}
+              />
+            </div> */}
             <div className="w-full flex flex-col gap-2 items-center justify-center bg-background p-[32px] border border-[#c9c9c9] mt-[32px]">
               <LoginForm />
             </div>